filebuffer: define core structs
Core structs for implementing a piece chain that also keep tracks of
lines.

In essences of lines are contiguous pieces of text. An insert or a
delete in the middle of a line creates a new line.

Possible optimization: a delete could be encoded with a special slice
structs that skips over some characters. (i.e chaining slices)

Another possible optimization: compression in the append buffer

Benefits/drawbacks:
 + easier parsing, processing with regex
 + easier line lookup and line navigation
 + easier blit to screen
 + less pointer indirection
 - mapped file needs to be traversed once at least
 - more pressure on the append buffer
 - more complex filebuffer internals